requirements.txt
Django==1.10.6
mysqlclient==1.3.10
South==0.8.4
django-registration==2.2
coverage==3.7.1
model_mommy==1.2
django-debug-toolbar==1.7
mock==1.0.1
django-model-utils==2.0
django-widget-tweaks==1.3
pip-tools==0.3.4
fabric==1.8.2
